+++ src/Model/ReturnLabelRequestBuilder.php
@@ -95,12 +95,13 @@ class ReturnLabelRequestBuilder implements ReturnLabelRequestBuilderInterface
     /**
      * Set package related data (optional).
      *
-     * @param float|null $weightInGrams Total weight of all included items (plus tare weight).
+     * @param int|null $weightInGrams Total weight of all included items (plus tare weight).
      * @param float|null $amount Total monetary value of all included items.
+     *
      * @return ReturnLabelRequestBuilderInterface
      */
     public function setPackageDetails(
-        float $weightInGrams = null,
+        int $weightInGrams = null,
         float $amount = null
     ): ReturnLabelRequestBuilderInterface {
         $this->data['package']['weight'] = $weightInGrams;
